requirements.txt
asttokens==2.4.1
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.10.0
dataclasses-json==0.6.5
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.13.1
fonttools==4.51.0
fsspec==2024.2.0
idna==2.10
intel-openmp==2021.4.0
ipython==8.24.0
ipywidgets==8.1.2
jedi==0.19.1
Jinja2==3.1.3
jupyter-bbox-widget==0.5.0
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.8.4
matplotlib-inline==0.1.7
mkl==2021.4.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
parso==0.8.4
pillow==10.3.0
prompt-toolkit==3.0.43
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
roboflow==1.1.28
scipy==1.13.0
# Editable Git install with no remote (segment_anything==1.0)
-e c:\users\dgn\desktop\annotation-sam\segment-anything-main
six==1.16.0
stack-data==0.6.3
supervision==0.20.0
sympy==1.12
tbb==2021.11.0
torch==2.3.0+cu121
torchaudio==2.3.0+cu121
torchvision==0.18.0+cu121
tqdm==4.66.2
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
widgetsnbextension==4.0.10
